requirements.txt
actionlib==1.14.0
angles==1.9.13
appdirs==1.4.4
av==11.0.0
brotlipy==0.7.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
certifi==2022.12.7
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.0.8
diagnostic-updater==1.11.0
distro==1.8.0
docker-pycreds==0.4.0
docutils==0.20.1
dynamic-reconfigure==1.7.3
easydict==1.13
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2024.2.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.11
GitPython==3.1.40
humanfriendly==10.0
importlib-resources==6.1.1
Jinja2==3.1.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
kiwisolver==1.4.5
lapx==0.5.5
MarkupSafe==2.1.5
matplotlib==3.7.4
message-filters==1.16.0
mistune==0.8.4
mkl-service==2.4.0
mpmath==1.3.0
nb-conda==2.2.1
networkx==3.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
onnxruntime==1.16.1
opencv-python==4.10.0.84
pandas==2.0.3
pathtools==0.1.2
Pillow==9.3.0
platformdirs==4.2.2
protobuf==5.28.0
psutil==5.9.8
py-cpuinfo==9.0.0
pycuda==2024.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytools==2024.1.5
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rosbag==1.16.0
rosclean==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslz4==1.16.0
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
scipy==1.10.1
seaborn==0.13.2
sensor-msgs==1.13.1
sentry-sdk==1.32.0
setproctitle==1.3.3
smclib==1.8.6
smmap==5.0.1
sympy==1.12
tellopy==0.6.0
tf==1.13.2
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
thop==0.1.1.post2209072238
topic-tools==1.16.0
torch==2.2.0
torchaudio==2.1.0+cu118
torchvision==0.17.0
tqdm==4.66.1
triton==2.2.0
typing_extensions==4.4.0
tzdata==2023.4
ultralytics==8.1.10
urllib3==1.26.13
wandb==0.16.1
webencodings==0.5.1
xacro==1.14.18
zipp==3.17.0
zstandard==0.22.0
zstd==1.5.5.1
